+++ Hazard-Utility/src/Buffer.h
@@ -5,13 +5,13 @@
 
 struct Buffer {
 
-	uint32_t Size = 0;
+	size_t Size = 0;
 	void* Data = nullptr;
 
 	Buffer() : Data(nullptr), Size(0) {};
-	Buffer(void* data, uint32_t size) : Data(data), Size(size) {};
+	Buffer(void* data, size_t size) : Data(data), Size(size) {};
 
-	void Allocate(uint32_t size)
+	void Allocate(size_t size)
 	{
 		delete Data;
 		Data = nullptr;
@@ -40,21 +40,21 @@ struct Buffer {
 			memset(Data, value, Size);
 	}
 	template<typename T>
-	T& Read(uint32_t offset = 0) {
+	T& Read(size_t offset = 0) {
 		return *(T*)((uint8_t*)Data + offset);
 	}
-	uint8_t* ReadBytes(uint32_t size, uint32_t offset) {
+	uint8_t* ReadBytes(size_t size, size_t offset) {
 		if (offset + size > Size) assert(false);
 		uint8_t* buffer = new uint8_t[size];
 
 		memcpy(buffer, (uint8_t*)Data + offset, size);
 		return buffer;
 	}
-	void Write(void* data, uint32_t size, uint32_t offset = 0) {
+	void Write(void* data, size_t size, size_t offset = 0) {
 		if (offset + size > Size) assert(false);
 		memcpy((uint8_t*)Data + offset, data, size);
 	}
-	void Write(const void* data, uint32_t size, uint32_t offset = 0) {
+	void Write(const void* data, size_t size, size_t offset = 0) {
 		if (offset + size > Size) assert(false);
 		memcpy((uint8_t*)Data + offset, data, size);
 	}
@@ -63,22 +63,22 @@ struct Buffer {
 	uint8_t operator[](int index) const { return ((uint8_t*)Data)[index]; }
 	template<typename T>
 	T* As() { return (T*)Data; }
-	inline uint32_t GetSize() { return Size; }
+	inline size_t GetSize() { return Size; }
 
-	static Buffer Copy(const void* data, uint32_t size, uint32_t offset = 0) {
+	static Buffer Copy(const void* data, size_t size, size_t offset = 0) {
 		Buffer buffer;
 		buffer.Allocate(size);
 		memcpy(buffer.Data, (uint8_t*)data + offset, size);
 		return buffer;
 	}
 	template<typename T>
-	static T Get(void* data, uint32_t startIndex = 0) {
+	static T Get(void* data, size_t startIndex = 0) {
 		T value;
 		memcpy(&value, (uint8_t*)data + startIndex, sizeof(T));
 		return value;
 	}
 	template<typename T>
-	static T GetRaw(void* data, uint32_t len, uint32_t startIndex = 0) {
+	static T GetRaw(void* data, size_t len, size_t startIndex = 0) {
 		T value;
 		memcpy(&value, data, len);
 		return value;
