+++ Hazard/src/Hazard/Rendering/RendererDrawList.h
@@ -11,10 +11,10 @@ namespace Hazard
 		glm::mat4 Transform;
 		Ref<HazardRenderer::VertexBuffer> VertexBuffer;
 		Ref<HazardRenderer::IndexBuffer> IndexBuffer;
-		uint32_t Count;
+		size_t Count;
 	};
 	struct PipelineData {
-		uint32_t Count;
+		size_t Count;
 	};
 
 	struct EnvironmentData
